refactor(deployments): move deployment-card files into own component subdirectory
related to openshiftio/openshift.io#1382
